Extracted classes/structs/enums from Common.cs into their own files.
