+++ SevenZip/CompressionMode.cs
@@ -1,0 +1,21 @@
+﻿#if UNMANAGED
+
+namespace SevenZip
+{
+    /// <summary>
+    /// Archive compression mode.
+    /// </summary>
+    public enum CompressionMode
+    {
+        /// <summary>
+        /// Create a new archive; overwrite the existing one.
+        /// </summary>
+        Create,
+        /// <summary>
+        /// Add data to the archive.
+        /// </summary>
+        Append,
+    }
+}
+
+#endif
